+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change Log
 ---
 
+- [Feature: Support custom rate limit reject response info](https://github.com/Tencent/spring-cloud-tencent/pull/154)
+- [Feature: Remove spring-javaformat-maven-plugin](https://github.com/Tencent/spring-cloud-tencent/pull/152)
+- [feat:optimize config server address](https://github.com/Tencent/spring-cloud-tencent/pull/150)
 - [feat:refactor loadbalancer module as a basic module for router and circuit breaker.](https://github.com/Tencent/spring-cloud-tencent/pull/166)